Add --no-prompt flag to avoid getting username and pass via prompt
This allows cookies to work as a lone authorization method whereas
previously it was required that you entered a username and password
every time. Now it will simply attempt to load all three of username,
password and cookies from environment variables and use what it gets.
The gist of it is that any credentials of the 3 total provided through
environment variables or saved in config must be correct, but you are
only required to give either username and pass OR cookies.